refactor: move postprocessSvg tests to lilypond/test/ (#474)
